fix: Psalm errors in PremiumService and ServerAdminService
PremiumService:
- getTwitchLiveMonthlyEurPrice: add ?? '4.99' fallback (settings->get returns ?string)
- subscribeTwitchLiveWithTokens/Paypal: use instanceof DateTimeImmutable check on
  getExpiresAt() before calling modify() (DateTimeInterface has no modify())
- getExpiresAt()->format(): use ?->format() ?? '' null-safe pattern (x2)
- canBookPosition/bookPosition: use DateTimeImmutable::createFromInterface() instead
  of (clone $startsAt)->modify() to avoid UndefinedInterfaceMethod
- psalm.xml: add PossiblyUnusedParam → info (param reserved for future use)

ServerAdminService:
- percent ternary: cast round() result to (int) to avoid float vs int strict operand

+++ src/Service/ServerAdminService.php
@@ -64,7 +64,7 @@ class ServerAdminService
             $p = preg_split('/\s+/', trim($mem['lines'][1]));
             $total = (int)($p[1] ?? 0);
             $used  = (int)($p[2] ?? 0);
-            $memData = ['total' => $total, 'used' => $used, 'percent' => $total > 0 ? round($used / $total * 100) : 0];
+            $memData = ['total' => $total, 'used' => $used, 'percent' => $total > 0 ? (int) round($used / $total * 100) : 0];
         }
 
         $diskData = ['total' => '-', 'used' => '-', 'percent' => 0];
